% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utm_crs.R
\name{get_circle}
\alias{get_circle}
\title{get_circle returns a circle of radius "radius" kilometers around lat/lon}
\usage{
get_circle(lat, lon, radius)
}
\arguments{
\item{radius}{}
}
\value{

}
\description{
get_circle returns a circle of radius "radius" kilometers around lat/lon
}
