% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utm_crs.R
\name{get_intersecting_geos}
\alias{get_intersecting_geos}
\title{get_intersecting_geos returns the CDs, CSDs and DAs that intersect with a geometry}
\usage{
get_intersecting_geos(geometry)
}
\arguments{
\item{geometry}{}
}
\value{

}
\description{
get_intersecting_geos returns the CDs, CSDs and DAs that intersect with a geometry
}
