% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utm_crs.R
\name{get_populations_from_polygon}
\alias{get_populations_from_polygon}
\title{get_populations_from_polygon   returns the canadian population within a polygon}
\usage{
get_populations_from_polygon(sf_tibble, use_cancensus = FALSE)
}
\arguments{
\item{sf_tibble}{}
}
\value{

}
\description{
get_populations_from_polygon   returns the canadian population within a polygon
}
