% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utm_crs.R
\name{get_utm_crs}
\alias{get_utm_crs}
\title{get_utm_crs returns the crs for a UTM in the northern hemisphere.  Example:    Montreal is in UTM zone 18, thus crs 32618}
\usage{
get_utm_crs(longitude)
}
\arguments{
\item{longitude}{}
}
\value{

}
\description{
get_utm_crs returns the crs for a UTM in the northern hemisphere.  Example:    Montreal is in UTM zone 18, thus crs 32618
}
